Move some config to pyproject.toml and delete config that are not useful anymore. Move conftest.py in the tests/ folder
