--- .coveragerc
@@ -1,6 +1,0 @@
-[run]
-source=fairworkflows
-branch=true
-
-[report]
-omit=fairworkflows/templates/*
