--- .flake8
@@ -1,9 +1,0 @@
-[flake8]
-
-exclude =
-    .git,
-    __pycache__,
-    fairworkflows/templates/,
-    fairworkflows/old,
-    venv,
-    env
