--- MANIFEST.in
@@ -1,6 +1,0 @@
-include README.md
-include CITATION.cff
-include fairworkflows/_version.py
-include requirements.txt
-include requirements_dev.txt
-include fairworkflows/resources/*
