--- pytest.ini
@@ -1,3 +1,0 @@
-# content of pytest.ini
-[pytest]
-testpaths = tests
