--- requirements.txt
@@ -1,9 +1,0 @@
-nanopub==1.2.7
-networkx~=2.5
-pytest
-pyyaml
-rdflib<6.0.0,>=5.0.0
-requests
-pyshacl>=0.14.1
-noodles==0.3.3
-Jinja2==2.11.3
