--- requirements_dev.txt
@@ -1,8 +1,0 @@
-coveralls
-flaky
-graphviz==0.14.1
-pytest
-pytest-cov
-recommonmark
-sphinx
-sphinx_rtd_theme
